+++ CHANGELOG.md
@@ -1,2 +1,12 @@
-**0.7.7**
- - toc bump
+# LobsterQuotes Changelog
+
+## 0.8.0
+
+**Total rewrite imminent?**
+
+* Background Alpha
+* Midnight pre-patch toc bump
+
+## 0.7.7
+
+toc bump